Removing sources of uncertainty from ClamAV
ClamAV is an anti-virus that scans directories by their
inode order, and thus being random in nature. This patch
does two things to address that: passes the files to scan
as a list (on the Makefile) and sorts the definition files
as they're read (readdb).

The result now should be identical on any machine/os/arch.


git-svn-id: https://llvm.org/svn/llvm-project@178660 91177308-0d34-0410-b5e6-96231b3b80d8