Improving lemon sort criteria makes test pass
As with other tests, the output of lemon was scrambled due to
its sorting criteria. Changing the comparison functions made it
more stable and now it's similar on Linux x86_64, ARM and Darwin.

The reference output is an MD5 of the real output because the
latter is 40M and diff has some trouble parsing it on resource
limited machines, not to mention skimming through a multi-MB
diff on the logs.


git-svn-id: https://llvm.org/svn/llvm-project@178810 91177308-0d34-0410-b5e6-96231b3b80d8