Add a new PlatformDarwinKernel for kernel debugging.  This Platform plugin will index the kext bundles on the local filesystem when created.  During a kernel debug session, when the DynamicLoader plugin needs to locate a kext by name like "com.apple.com.apple.filesystems.autofs", the Platform can quickly look for a UUID match in those kernel debug kit directories it previously indexed.
I'm still working on profiling the performance impact of the inital
kext bundle scan; there will likely need to be a switch to enable
or disable this plugin's scan.

This only affects Mac kernel debugging and the code is only built
on Apple systems because of some use of low-level CoreFoundation
to parse plists.

<rdar://problem/13503583> 



git-svn-id: https://llvm.org/svn/llvm-project@178827 91177308-0d34-0410-b5e6-96231b3b80d8