Fix office-ispell big-endian reference output
As Daniel has mentioned, this test is completely broken (on both big and little
endian systems, it just errors out). On top of that, there seems to be some
endian sensitivity in the code that checks the file magic. As a result, on big
endian systems, the error message contains the full path of the input file
(which makes it impossible to create a reference output).

Here I've slightly modified the test so that it only prints the base path name
on big endian systems (so that I can create a reference output), and I've added
that reference output. This change can be reverted once the test is actually
fixed to do something useful.


git-svn-id: https://llvm.org/svn/llvm-project@181258 91177308-0d34-0410-b5e6-96231b3b80d8