cpp11-migrate: Add headers to file override information
File override structures now contain per-source overrides for headers.

For now, modified headers are written to disk when the Migrator is done.
This is only temporary behaviour since we can expect headers to be
changed by migrating multiple source files. The changes need to be
merged after all migrations are complete.



git-svn-id: https://llvm.org/svn/llvm-project@183917 91177308-0d34-0410-b5e6-96231b3b80d8