Remove the sysroot (or isysroot) restriction from the GCDAProfile.c runtime in the Makefile build system as well. Sorry for the temporary fallout, it took me a while to find these bits.
Bill, I'm not at all confident of the change for Darwin and iOS, but as
discussed we're completely blocked on fixing this. Anyways, please
review and let me know if this will work for you guys. If necessary,
I can work with you to rig up an errno.h stub for Darwin, but I expect
that to be... moderately challenging.

git-svn-id: https://llvm.org/svn/llvm-project@184805 91177308-0d34-0410-b5e6-96231b3b80d8