BlockGenerator: Simplify the old value searching code.
Orignally, we first test if a ValueMap contains a Value, and than use the
index operator to get the corresponding new value. This requires the ValueMap
to lookup the key (i.e. the old value) twice.

Now, we directly use the "lookup" function provided by DenseMap to implement
the same functionality.

git-svn-id: https://llvm.org/svn/llvm-project@185260 91177308-0d34-0410-b5e6-96231b3b80d8