PR16480: Reimplement token-caching for constructor initializer lists. This previously didn't work if a mem-initializer-id had a template argument which contained parentheses or braces.
We now implement a simple rule: just look for a ') {' or '} {' that is not
nested. The '{' is assumed to start the function-body. There are still two
cases which we misparse, where the ') {' comes from a compound literal or
from a lambda. The former case is not valid C++, and the latter will probably
not be valid C++ once DR1607 is resolved, so these seem to be of low value,
and we do not regress on them with this change. EDG and g++ also misparse
both of these cases.


git-svn-id: https://llvm.org/svn/llvm-project@185598 91177308-0d34-0410-b5e6-96231b3b80d8